glaucoma_backend_fixed/glaucoma_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
psycopg2-binary
alembic
Pillow
numpy
